refactor: atom 각각을 관련 있는 모듈에서 관리하도록 분리
