+++ src/recoil/atoms/common.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { ToastProps } from '../../components/common/Toast/Toast';
+
+export const toastState = atom<ToastProps[]>({
+  key: 'toastState',
+  default: [],
+});
